AMC/requirements.txt
APScheduler>=3.10.4
playwright>=1.40.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
